# packages/contracts/openapi/openapi.json
{
  "components": {
    "parameters": {
...
      "Cursor": {
        "in": "query",
        "name": "cursor",
        "required": false,
        "schema": {
          "type": [
            "string",
            "null"
          ]
        }
      },
      "IdempotencyKey": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 8,
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Sprint2Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (401)",
        "x-error-codes": [
          "AUTH_REQUIRED"
        ],
        "x-http-status": 401
      },
      "Sprint2Error403": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (403)",
        "x-error-codes": [
          "FORBIDDEN",
          "PERMISSION_CHANGED"
        ],
        "x-http-status": 403
      },
      "Sprint2Error404": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (404)",
        "x-error-codes": [
          "RESOURCE_NOT_FOUND"
        ],
        "x-http-status": 404
      },
      "Sprint2Error409": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (409)",
        "x-error-codes": [
          "VERSION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "CLARIFICATION_ROUND_LIMIT_REACHED",
          "DEEP_CONFIRMATION_REQUIRED",
          "AI_TASK_NOT_CANCELLABLE",
          "AI_TASK_NOT_RETRYABLE",
          "RESULT_NOT_READY",
          "RESULT_QUALITY_BLOCKED",
          "STALE_TARGET",
          "TRACEABILITY_INCOMPLETE",
          "CLARIFICATION_ROUND_INVALID"
        ],
        "x-http-status": 409
      },
      "Sprint2Error422": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (422)",
        "x-error-codes": [
          "VALIDATION_ERROR",
          "CLARIFICATION_MODE_REASON_REQUIRED",
          "RISK_ACCEPTANCE_INVALID"
        ],
        "x-http-status": 422
      },
      "Sprint2Error429": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (429)",
        "x-error-codes": [
          "AI_QUOTA_EXCEEDED"
        ],
        "x-http-status": 429
      },
      "Sprint2Error503": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Sprint2ErrorResponse"
            }
          }
        },
        "description": "Sprint 2 error (503)",
        "x-error-codes": [
          "QUEUE_UNAVAILABLE",
          "DEPENDENCY_UNAVAILABLE"
        ],
        "x-http-status": 503
      },
...
    },
    "schemas": {
...
      "AiResultStatus": {
        "enum": [
          "ready",
          "partial_result",
          "quality_blocked",
          "failed",
          "expired",
          "stale_target"
        ],
        "type": "string"
      },
...
      "AiTaskListData": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AiTaskSummary"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "type": "object"
      },
      "AiTaskListResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AiTaskListData"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "trace_id"
        ],
        "type": "object"
      },
      "AiTaskResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AiTaskSummary"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "trace_id"
        ],
        "type": "object"
      },
      "AiTaskResultRef": {
        "additionalProperties": false,
        "properties": {
          "result_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AiResultStatus"
          },
          "target_snapshot_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          }
        },
        "required": [
          "result_id",
          "status",
          "target_snapshot_hash"
        ],
        "type": "object"
      },
      "AiTaskStatus": {
        "enum": [
          "prechecking",
          "blocked",
          "queued",
          "preparing",
          "generating",
          "checking",
          "ready",
          "partial_result",
          "quality_blocked",
          "cancel_requested",
          "cancelled",
          "failed",
          "expired",
          "stale_target"
        ],
        "type": "string"
      },
      "AiTaskSummary": {
        "additionalProperties": false,
        "properties": {
          "capability_summary": {
            "additionalProperties": true,
            "type": "object"
          },
          "created_by_user_id": {
            "type": "string"
          },
          "events_url": {
            "format": "uri-reference",
            "type": "string"
          },
          "missing_items": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "poll_url": {
            "format": "uri-reference",
            "type": "string"
          },
          "queued_at": {
            "format": "date-time",
            "type": "string"
          },
          "result_refs": {
            "items": {
              "$ref": "#/components/schemas/AiTaskResultRef"
            },
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AiTaskStatus"
          },
          "target_snapshot_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "task_public_id": {
            "format": "uuid",
            "type": "string"
          },
          "task_type": {
            "const": "requirement.clarify",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "task_public_id",
          "status",
          "task_type",
          "created_by_user_id",
          "queued_at",
          "target_snapshot_hash",
          "capability_summary",
          "missing_items",
          "result_refs",
          "events_url",
          "poll_url"
        ],
        "type": "object"
      },
...
      "ClarificationMode": {
        "enum": [
          "auto",
          "standard",
          "deep",
          "skip"
        ],
        "type": "string"
      },
...
      "ClarificationTaskInput": {
        "additionalProperties": false,
        "properties": {
          "continue_deep_confirmed": {
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/ClarificationMode"
          },
          "round_no": {
            "maximum": 5,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "mode",
          "round_no",
          "continue_deep_confirmed"
        ],
        "type": "object"
      },
...
      "CreateAiTaskRequest": {
        "additionalProperties": false,
        "properties": {
          "capability_selection": {
            "additionalProperties": true,
            "type": [
              "object",
              "null"
            ]
          },
          "risk_acceptances": {
            "items": {
              "$ref": "#/components/schemas/RiskAcceptance"
            },
            "type": "array"
          },
          "source_ref_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array",
            "uniqueItems": true
          },
          "target": {
            "$ref": "#/components/schemas/TaskTarget"
          },
          "task_type": {
            "const": "requirement.clarify",
            "type": "string"
          },
          "user_instruction": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "task_type",
          "target",
          "source_ref_ids"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "RequirementClarifyTaskEnvelope": {
        "additionalProperties": false,
        "properties": {
          "capability_selection": {
            "additionalProperties": true,
            "type": [
              "object",
              "null"
            ]
          },
          "command_id": {
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/ClarificationTaskInput"
          },
          "module": {
            "const": "product_design",
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "project_version_id": {
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "type": "string"
          },
          "risk_acceptances": {
            "items": {
              "$ref": "#/components/schemas/RiskAcceptance"
            },
            "type": "array"
          },
          "schema_version": {
            "const": "0.2.0",
            "type": "string"
          },
          "source_ref_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array",
            "uniqueItems": true
          },
          "target": {
            "$ref": "#/components/schemas/TaskTarget"
          },
          "target_snapshot_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "task_public_id": {
            "type": "string"
          },
          "task_type": {
            "const": "requirement.clarify",
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "task_public_id",
          "user_id",
          "project_id",
          "project_version_id",
          "module",
          "task_type",
          "target",
          "target_snapshot_hash",
          "source_ref_ids",
          "capability_selection",
          "risk_acceptances",
          "command_id",
          "trace_id",
          "requested_at",
          "input"
        ],
        "type": "object"
      },
...
      "RiskAcceptance": {
        "additionalProperties": false,
        "properties": {
          "impact": {
            "$ref": "#/components/schemas/RiskImpact"
          },
          "missing_item_code": {
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "missing_item_code",
          "impact",
          "reason"
        ],
        "type": "object"
      },
      "RiskImpact": {
        "enum": [
          "low",
          "medium"
        ],
        "type": "string"
      },
...
      "Sprint2ErrorCode": {
        "enum": [
          "AI_QUOTA_EXCEEDED",
          "AI_TASK_NOT_CANCELLABLE",
          "AI_TASK_NOT_RETRYABLE",
          "AUTH_REQUIRED",
          "CLARIFICATION_MODE_REASON_REQUIRED",
          "CLARIFICATION_ROUND_INVALID",
          "CLARIFICATION_ROUND_LIMIT_REACHED",
          "DEEP_CONFIRMATION_REQUIRED",
          "DEPENDENCY_UNAVAILABLE",
          "FORBIDDEN",
          "IDEMPOTENCY_CONFLICT",
          "PERMISSION_CHANGED",
          "QUEUE_UNAVAILABLE",
          "RESOURCE_NOT_FOUND",
          "RESULT_NOT_READY",
          "RESULT_QUALITY_BLOCKED",
          "RISK_ACCEPTANCE_INVALID",
          "STALE_TARGET",
          "TRACEABILITY_INCOMPLETE",
          "VALIDATION_ERROR",
          "VERSION_CONFLICT"
        ],
        "type": "string"
      },
      "Sprint2ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/Sprint2ErrorCode"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "trace_id"
        ],
        "type": "object"
      },
...
      "TaskTarget": {
        "additionalProperties": false,
        "properties": {
          "object_id": {
            "type": "string"
          },
          "object_type": {
            "const": "requirement",
            "type": "string"
          },
          "object_version_id": {
            "type": "string"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "object_version_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "refreshCookie": {
        "description": "Secure, HttpOnly, SameSite refresh cookie; default lifetime seven days.",
        "in": "cookie",
        "name": "refresh_token",
        "type": "apiKey"
      },
      "serviceBearerAuth": {
        "bearerFormat": "JWT",
        "description": "Internal HS256 service JWT; fixed issuer/audience/scope and short lifetime.",
        "scheme": "bearer",
        "type": "http",
        "x-ai-caller-subject": "ai-api",
        "x-required-audience": "business-api",
        "x-required-issuers": [
          "ai-api",
          "monitoring"
        ],
        "x-required-scope": "health"
      }
    }
  },
  "info": {
    "description": "Sprint 2 Requirement/Baseline R4 implementation candidate authorized by PORTFOLIO-P1-RUNTIME-01. It remains pending Review and the shared real-runtime Gate; persistence adapters and FLOW remain disabled.",
    "summary": "Sprint 2 Requirement/Baseline R4 implementation candidate",
    "title": "AI Product Design and Validation API",
    "version": "0.1.0-dev",
    "x-contract-reviewers": [
      "frontend",
      "ai-data",
      "review"
    ],
    "x-contract-status": "candidate"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/tasks": {
      "get": {
        "operationId": "listAiTasks",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiTaskListResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "ai-tasks"
        ],
        "x-command-semantics": {
          "actor": "project_member_with_target_read_permission",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-r4-decisions": [
          "permission"
        ]
      },
      "post": {
        "description": "Creates one traceable requirement.clarify Task; it does not call a Provider synchronously.",
        "operationId": "createAiTask",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAiTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiTaskResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "ai-tasks"
        ],
        "x-command-envelope": {
          "derived_input": "Business API derives and validates input from target Requirement Version; frontend input is not trusted",
          "excluded_from_public_request": [
            "user_id",
            "command_id",
            "trace_id",
            "requested_at",
            "project_id",
            "project_version_id",
            "module",
            "input"
          ],
          "ref": "#/components/schemas/RequirementClarifyTaskEnvelope"
        },
        "x-command-semantics": {
          "actor": "owner",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": true
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner"
          ]
        },
        "x-r4-decisions": [
          "permission",
          "error_codes",
          "task_acceptance",
          "transaction",
          "event"
        ],
        "x-transaction": {
          "atomic": true,
          "audit_required": true,
          "boundary": "task+idempotency+audit; queue_acceptance_not_false_success",
          "outbox_required": true,
          "task_creation_decoupled": false
        },
        "x-transaction-boundary": "task+idempotency+audit; queue_acceptance_not_false_success"
      }
    },
    "/api/v1/ai/tasks/{task_id}": {
      "get": {
        "operationId": "getAiTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiTaskResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Sprint2Error401"
          },
          "403": {
            "$ref": "#/components/responses/Sprint2Error403"
          },
          "404": {
            "$ref": "#/components/responses/Sprint2Error404"
          },
          "409": {
            "$ref": "#/components/responses/Sprint2Error409"
          },
          "422": {
            "$ref": "#/components/responses/Sprint2Error422"
          },
          "429": {
            "$ref": "#/components/responses/Sprint2Error429"
          },
          "503": {
            "$ref": "#/components/responses/Sprint2Error503"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "ai-tasks"
        ],
        "x-command-semantics": {
          "actor": "project_member_with_target_read_permission",
          "admin_bypass": false,
          "queue_unavailable_no_false_queued": false
        },
        "x-contract-phase": "sprint-2-p1-runtime-r4-candidate",
        "x-error-http-mapping": {
          "AI_QUOTA_EXCEEDED": 429,
          "AI_TASK_NOT_CANCELLABLE": 409,
          "AI_TASK_NOT_RETRYABLE": 409,
          "AUTH_REQUIRED": 401,
          "CLARIFICATION_MODE_REASON_REQUIRED": 422,
          "CLARIFICATION_ROUND_INVALID": 409,
          "CLARIFICATION_ROUND_LIMIT_REACHED": 409,
          "DEEP_CONFIRMATION_REQUIRED": 409,
          "DEPENDENCY_UNAVAILABLE": 503,
          "FORBIDDEN": 403,
          "IDEMPOTENCY_CONFLICT": 409,
          "PERMISSION_CHANGED": 403,
          "QUEUE_UNAVAILABLE": 503,
          "RESOURCE_NOT_FOUND": 404,
          "RESULT_NOT_READY": 409,
          "RESULT_QUALITY_BLOCKED": 409,
          "RISK_ACCEPTANCE_INVALID": 422,
          "STALE_TARGET": 409,
          "TRACEABILITY_INCOMPLETE": 409,
          "VALIDATION_ERROR": 422,
          "VERSION_CONFLICT": 409
        },
        "x-implementation-status": "implemented-p1-runtime-r4-candidate",
        "x-permission": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-permissions": {
          "admin_bypass": false,
          "allowed_project_roles": [
            "owner",
            "reviewer",
            "implementer",
            "tester"
          ]
        },
        "x-r4-decisions": [
          "permission"
        ]
      }
    },
...
  },
...
}